fixed(redux): reorganizacion de carpetas
