+++ client/src/Pages/Register.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
 import { useState } from "react";
-import postUser from "../redux/actions";
+import postUser from "../redux/actions/actionsPetitions";
 import logo from '../assets/logo.png'
 import '../styles/register.css'
