+++ client/src/Pages/ShoppingCart.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+
+export default function ShoppingCart(){
+    return(
+        <div>CARRITO DE COMPRAS EN PROCESO...</div>
+    )
+}
