+++ client/src/Pages/home.jsx
@@ -4,7 +4,7 @@ import Carrusel from '../componets/Carrusel';
 import CardProductContainer from '../componets/CardProductContainer';
 import { Link } from "react-router-dom";
 import {useDispatch} from "react-redux"
-import { getAllProducts } from "../redux/actions";
+import { getAllProducts } from "../redux/actions/actionsPetitions";
 
 export default function Home(){
     const dispatch = useDispatch()
