+++ client/src/componets/Filters.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useDispatch, useSelector } from 'react-redux'
-import { getAllProducts, orderByPrice, filterCategory } from "../redux/actions";
+import { getAllProducts, orderByPrice, filterCategory } from "../redux/actions/actionsPetitions";
 
 
 export default function Filters({ setOption, setPage }) {
