+++ client/src/redux/actions/actionShoppingCart.js
@@ -1,0 +1,30 @@
+import axios from "axios";
+
+
+export function shoppingCart(){
+    return {
+        type: 'ADD_TO_CART',
+        payload
+    }
+}
+
+export function shoppingCart(){
+    return {
+        type: 'REMOVE_ONE_CART',
+        payload
+    }
+}
+
+export function shoppingCart(){
+    return {
+        type: 'REMOVE_ALL_CART',
+        payload
+    }
+}
+
+export function shoppingCart(){
+    return {
+        type: 'CLEAR_CART',
+        payload
+    }
+}
