+++ client/src/redux/actions/actionsPetitions.js
@@ -45,3 +45,4 @@ export function filterCategory(payload) {
         payload
     }
 }
+
