+++ client/src/redux/store.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from "redux-devtools-extension";
 import thunk from 'redux-thunk';
-import rootReducer from './reducer';
+import rootReducer from './reducers/reducerPetitions';
 
 
 const store = createStore(rootReducer,
